Add files via upload
Updated code as of Tuesday 25/9/18
Extensive comments to explain and justify what is going on.
It appears to work correctly when compared to what is expected from real life and Ben's old code, however I think i should be dividing the fft() by n and multiplying by 2 when it is cut in half. It  does appear to be working correctly though despite this confusion!